# openapiconfig-privsense.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "PrivSense API",
      "description": "API for detecting personally identifiable information (PII) in relational databases",
      "contact": {
        "name": "PrivSense Team",
        "url": "https://www.privsense.com",
        "email": "support@privsense.com"
      },
      "license": {
        "name": "Proprietary",
        "url": "https://www.privsense.com/license"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/privsense",
        "description": "Default server URL"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
...
    "paths": {
...
      "/api/v1/scan-templates/{id}": {
        "get": {
          "tags": [
            "Scan Templates"
          ],
          "summary": "Get template",
          "description": "Returns the specified scan template",
          "operationId": "getTemplate",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Template ID",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "404": {
              "description": "Template not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            },
            "200": {
              "description": "Template retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Scan Templates"
          ],
          "summary": "Update scan template",
          "description": "Updates an existing scan template",
          "operationId": "updateTemplate",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Template ID",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanTemplateDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "400": {
              "description": "Invalid template data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            },
            "200": {
              "description": "Template updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            },
            "404": {
              "description": "Template or connection not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Scan Templates"
          ],
          "summary": "Delete template",
          "description": "Deletes the specified scan template",
          "operationId": "deleteTemplate",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Template ID",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "404": {
              "description": "Template not found"
            },
            "204": {
              "description": "Template deleted successfully"
            }
          }
        }
      },
...
      "/api/v1/scans": {
        "get": {
          "tags": [
            "PII Scans"
          ],
          "summary": "List all scans",
          "description": "Returns a paginated list of all scan jobs in the system (Admins see all, API users see only their scans)",
          "operationId": "getAllScans",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Page number (0-based)",
              "required": false,
              "schema": {
                "type": "string",
                "default": "0"
              }
            },
            {
              "name": "size",
              "in": "query",
              "description": "Number of items per page",
              "required": false,
              "schema": {
                "type": "string",
                "default": "20"
              }
            },
            {
              "name": "status",
              "in": "query",
              "description": "Filter by scan status (completed, pending, failed)",
              "required": false,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "connectionId",
              "in": "query",
              "description": "Filter by database connection ID",
              "required": false,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of scan jobs retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponse"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "PII Scans"
          ],
          "summary": "Start a new scan",
          "description": "Initiates a new database scan to detect PII in the database (Admin only)",
          "operationId": "startScan",
          "requestBody": {
            "description": "Scan configuration parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Scan job created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Database connection not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/scan-templates": {
        "get": {
          "tags": [
            "Scan Templates"
          ],
          "summary": "Get all templates",
          "description": "Returns all available scan templates",
          "operationId": "getAllTemplates",
          "parameters": [
            {
              "name": "connectionId",
              "in": "query",
              "description": "Filter by database connection ID",
              "required": false,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of templates retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Scan Templates"
          ],
          "summary": "Create scan template",
          "description": "Saves a new scan template for future use",
          "operationId": "createTemplate",
          "requestBody": {
            "description": "Template data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanTemplateDTO"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Template created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            },
            "404": {
              "description": "Connection not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid template data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanTemplateDTO"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/scan-templates/{id}/execute": {
        "post": {
          "tags": [
            "Scan Templates"
          ],
          "summary": "Execute scan from template",
          "description": "Starts a new scan using the specified template configuration",
          "operationId": "executeScanFromTemplate",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Template ID",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            },
            {
              "name": "connectionId",
              "in": "query",
              "description": "Override the connection ID from the template",
              "required": false,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Scan job created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Template or connection not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/scans/{jobId}": {
        "get": {
          "tags": [
            "PII Scans"
          ],
          "summary": "Get scan status",
          "description": "Returns the current status of a running scan job",
          "operationId": "getScanStatus",
          "parameters": [
            {
              "name": "jobId",
              "in": "path",
              "description": "ID of the scan job",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Scan status retrieved successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Scan job not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ScanJobResponse"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "PII Scans"
          ],
          "summary": "Cancel scan",
          "description": "Cancels an in-progress scan job (Admin only)",
          "operationId": "cancelScan",
          "parameters": [
            {
              "name": "jobId",
              "in": "path",
              "description": "ID of the scan job to cancel",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "409": {
              "description": "Scan already completed or failed"
            },
            "404": {
              "description": "Scan job not found"
            },
            "204": {
              "description": "Scan cancelled successfully"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "DetectionConfigDTO": {
          "type": "object",
          "properties": {
            "heuristicThreshold": {
              "type": "number",
              "description": "Threshold for heuristic-based detection",
              "format": "double",
              "example": 0.7
            },
            "regexThreshold": {
              "type": "number",
              "description": "Threshold for regex-based detection",
              "format": "double",
              "example": 0.8
            },
            "nerThreshold": {
              "type": "number",
              "description": "Threshold for NER-based detection",
              "format": "double",
              "example": 0.6
            },
            "reportingThreshold": {
              "type": "number",
              "description": "Minimum threshold for reporting PII findings",
              "format": "double",
              "example": 0.5
            },
            "stopPipelineOnHighConfidence": {
              "type": "boolean",
              "description": "Whether to stop detection pipeline on high confidence matches"
            },
            "entropyEnabled": {
              "type": "boolean",
              "description": "Whether entropy-based detection is enabled"
            }
          },
          "description": "PII detection configuration settings"
        },
        "SamplingConfigDTO": {
          "type": "object",
          "properties": {
            "defaultSize": {
              "type": "integer",
              "description": "Default sample size",
              "format": "int32",
              "example": 100
            },
            "maxConcurrentDbQueries": {
              "type": "integer",
              "description": "Maximum concurrent database queries",
              "format": "int32",
              "example": 10
            },
            "entropyCalculationEnabled": {
              "type": "boolean",
              "description": "Whether entropy calculation is enabled during sampling"
            },
            "defaultMethod": {
              "type": "string",
              "description": "Default sampling method",
              "example": "RANDOM"
            }
          },
          "description": "Data sampling configuration settings"
        },
        "ScanTemplateDTO": {
          "required": [
            "connectionId",
            "name"
          ],
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "name": {
              "maxLength": 255,
              "minLength": 0,
              "type": "string"
            },
            "description": {
              "maxLength": 1000,
              "minLength": 0,
              "type": "string"
            },
            "createdBy": {
              "type": "string"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            },
            "connectionId": {
              "type": "string",
              "format": "uuid"
            },
            "targetTables": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "samplingConfig": {
              "$ref": "#/components/schemas/SamplingConfigDTO"
            },
            "detectionConfig": {
              "$ref": "#/components/schemas/DetectionConfigDTO"
            },
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            }
          }
        },
...
        "ScanRequest": {
          "required": [
            "connectionId",
            "detectionConfig",
            "samplingConfig"
          ],
          "type": "object",
          "properties": {
            "connectionId": {
              "type": "string",
              "format": "uuid"
            },
            "targetTables": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "samplingConfig": {
              "$ref": "#/components/schemas/SamplingConfigDTO"
            },
            "detectionConfig": {
              "$ref": "#/components/schemas/DetectionConfigDTO"
            }
          }
        },
        "Link": {
          "type": "object",
          "properties": {
            "href": {
              "type": "string"
            },
            "hreflang": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "deprecation": {
              "type": "string"
            },
            "profile": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "templated": {
              "type": "boolean"
            }
          }
        },
        "ScanJobResponse": {
          "type": "object",
          "properties": {
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "links": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Link"
              }
            },
            "jobId": {
              "type": "string",
              "format": "uuid"
            },
            "connectionId": {
              "type": "string",
              "format": "uuid"
            },
            "status": {
              "type": "string"
            },
            "currentOperation": {
              "type": "string"
            },
            "startTime": {
              "type": "string",
              "format": "date-time"
            },
            "endTime": {
              "type": "string",
              "format": "date-time"
            },
            "lastUpdateTime": {
              "type": "string",
              "format": "date-time"
            },
            "progress": {
              "type": "integer",
              "format": "int32"
            },
            "completed": {
              "type": "boolean"
            },
            "failed": {
              "type": "boolean"
            },
            "errorMessage": {
              "type": "string"
            },
            "databaseName": {
              "type": "string"
            },
            "databaseProductName": {
              "type": "string"
            },
            "totalColumnsScanned": {
              "type": "integer",
              "format": "int32"
            },
            "totalPiiColumnsFound": {
              "type": "integer",
              "format": "int32"
            },
            "success": {
              "type": "boolean"
            }
          }
        },
...
        "PageResponse": {
          "type": "object",
          "properties": {
            "meta": {
              "type": "object",
              "additionalProperties": {
                "type": "object"
              }
            },
            "links": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Link"
              }
            },
            "content": {
              "type": "array",
              "items": {
                "type": "object"
              }
            },
            "pageNumber": {
              "type": "integer",
              "format": "int32"
            },
            "pageSize": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            },
            "hasNext": {
              "type": "boolean"
            },
            "hasPrevious": {
              "type": "boolean"
            },
            "success": {
              "type": "boolean"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "Enter JWT token with Bearer prefix. Example: Bearer eyJhbGciOiJIUzI1NiIsIn...",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }